linux: Add linux-sitec tree for linux-imx
Remove linux-imx recipes and patches and add linux-sitec. Also add S4 specific patches
to the kernel tree of the layer. Now the meta-sitec-score layers builds a kernel which can
be used on S Core only and on the S4.
The linux recipe is now named like other recipes in the community (linux-company name).

Ticket #1687
Ticket #1688