u-boot-imx: Add scripts for mfg boot
Add a script for flashing the SPI NOR with a valid u-boot image. The u-boot image
is loaded from RAM. The script depends on the mfg tool which needs to load the
u-boot image into the defined RAM location.
Also add script for writing the fuses.

Remove some unused defines.

Ticket #1763